pyproject.toml
[project]
name = "langchain-foundation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.4",
    "bs4>=0.0.2",
    "chromadb>=1.0.15",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.116.0",
    "ipykernel>=6.29.5",
    "ipython>=8.37.0",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-groq>=0.3.5",
    "langchain-huggingface>=0.3.0",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.25",
    "langchain-tavily>=0.2.11",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.6.3",
    "langgraph-cli[inmem]>=0.4.0",
    "langserve>=0.3.1",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "sse-starlette>=3.0.2",
    "streamlit>=1.46.1",
    "tiktoken>=0.9.0",
    "wikipedia>=1.4.0",
]
